Tune Android SDK 5.2.1

+++ sdk-android/TuneMarketingConsoleSDK/src/main/java/com/tune/ma/analytics/model/event/TuneAnalyticsEventBase.java
@@ -57,10 +57,10 @@ public abstract class TuneAnalyticsEventBase {
     private List<TuneAnalyticsVariable> profile;
 
     private double sessionTime = -1;
-    private double timeStamp = -1;
+    private long timeStamp = -1L;
 
     public TuneAnalyticsEventBase() {
-        this.timeStamp = System.currentTimeMillis() / 1000.0;
+        this.timeStamp = System.currentTimeMillis() / 1000L;
 
         if (TuneManager.getInstance() != null) {
             if (TuneManager.getInstance().getProfileManager() != null) {
@@ -146,7 +146,7 @@ public abstract class TuneAnalyticsEventBase {
         this.tags.addAll(tags);
     }
 
-    public double getTimeStamp() {
+    public long getTimeStamp() {
         return timeStamp;
     }
